Create Print triplets - 2020  a1 87
Write a static method that reduces as a whole array array parameter that is complete, positive and negative.
The method has to find a triad of numbers in the array (not necessarily adjacent to each other)
Is the maximum. The method will print these three numbers and return the product
The maximum. No matter the print order.
Please note - you must print the three numbers spaced without extra text.
One can assume that the array length is simple 3.